(1350) ajout mixin popup and thumbnail
